+++ elements/timecontrol/src/main.js
@@ -29,9 +29,6 @@ export class EOxTimeControl extends LitElement {
        */
       controlValues: { type: Array, attribute: "control-values" },
 
-      /**
-       * The query selector for the map
-       */
       for: { type: String },
 
       /**
@@ -59,6 +56,7 @@ export class EOxTimeControl extends LitElement {
       _controlSource: { state: true },
       _isAnimationPlaying: { state: true },
       _newStepIndex: { state: true },
+      _eoxMap: { state: true },
       unstyled: { type: Boolean },
     };
   }
@@ -75,12 +73,19 @@ export class EOxTimeControl extends LitElement {
     this.disablePlay = false;
     /** @type {boolean} */
     this.slider = false;
-    /** @type {string|HTMLElement} */
-    this.for = "";
+    /**
+     * Query selector of an `eox-map` (`String`, passed as an attribute or property)
+     * or an `eox-map` DOM element (`HTMLElement`, passed as property)
+     *
+     * @type {String|HTMLElement}
+     */
+    this.for = "eox-map";
     /** @type {string} */
     this.layer = "";
     /** @type {string | undefined} */
     this.controlProperty = undefined;
+    /** @type {HTMLElement |undefined} */
+    this._eoxMap = undefined;
   }
 
   /**
@@ -141,6 +146,41 @@ export class EOxTimeControl extends LitElement {
     } else {
       console.error(`Unable to find step "${step}" in available times!`);
     }
+  }
+
+  /**
+   * initializes the EOxMap instance
+   * And stores it in the private property #eoxMap.
+   */
+  firstUpdated() {
+    this.updateMap();
+  }
+
+  updated(changedProperties) {
+    if (changedProperties.has("for")) {
+      this.updateMap();
+    }
+  }
+
+  updateMap() {
+    const foundElement = getElement(this.for);
+
+    if (foundElement) {
+      const EoxMap = /** @type {import("@eox/map/main").EOxMap} */ (
+        foundElement
+      );
+      this.eoxMap = EoxMap;
+    }
+  }
+
+  get eoxMap() {
+    return this._eoxMap;
+  }
+
+  set eoxMap(value) {
+    const oldValue = this._eoxMap;
+    this._eoxMap = value;
+    this.requestUpdate("eoxMap", oldValue);
   }
 
   /**